Split elf tests
